# src/schema/api.yml
swagger: '2.0'
info:
  version: '1.0'
  title: 'IXO'
  description: ''
host: block-sync-pandora.ixo.world/transactions
schemes:
  - https
  - http
paths:
  '/':
    get:
      description: Ping
      summary: ping
      operationId: ping
      tags: [ dippersync ]
      responses:
        200:
          description: Api
  '/listTransactionsByAddr/{addr}':
    get:
      description: Returns a list of sends and receives for a given address
      operationId: listTransactionsByAddr
      tags: [ transactions ]
      parameters:
        - in: path
          name: addr
          type: string
          required: true
          description: ixo address
      responses:
        200:
          description: List of send and receive transactions 
          schema:
            $ref: '#/definitions/Transactions'
  '/listTransactionsByAddrByAsset/{addr}/{asset}':
    get:
      description: Returns a list of sends and receives for a given address and a given denom
      operationId: listTransactionsByAddrByAsset
      tags: [ transactions ]
      parameters:
        - in: path
          name: addr
          type: string
          required: true
          description: ixo address
        - in: path
          name: asset
          type: string
          required: true
          description: token denom  
      responses:
        200:
          description: List of send and receive transactions 
          schema:
            $ref: '#/definitions/Transactions' 
  '/listTransactionsByBondDid/{did}':
    get:
      description: Returns a list buys and sells for a given bond did
      operationId: listTransactionsByBondDid
      tags: [ transactions ]
      parameters:
        - in: path
          name: did
          type: string
          required: true
          description: bond did
      responses:
        200:
          description: List of buy and sell transactions
          schema:
            $ref: '#/definitions/Transactions'                    
  
definitions:
  Transactions:
    type: array
    items:
      $ref: '#/definitions/Transaction'
  Transaction:
    type: object
    properties:
      txhash: 
        type: string
      height: 
        type: number
      processed:
        type: boolean
      tx: 
        $ref: '#/definitions/Tx'
      tx_response: 
        $ref: '#/definitions/TxResponse'
  Tx:
    type: object
    properties:
      body:
        type: object
        properties:
          messages:
            type: array
            items:
              $ref: '#/definitions/TypeValue'
          memo: 
            type: string
          timeout_height:
            type: string
          extension_options: 
            type: array
            items: 
              $ref: '#/definitions/TypeValue'
          non_critical_extension_options:
            type: array
            items: 
              $ref: '#/definitions/TypeValue'
      auth_info:
        type: object
        properties:
          signer_infos:
            type: array
            items:
              $ref: '#/definitions/Signer' 
          fee:
            type: object
            properties:
              amount:
                type: array
                items: 
                  $ref: '#/definitions/Amount'
              gas_limit:
                type: string
              payer:
                type: string
              granter:
                type: string
      signatures:
        type: array
        items: 
          type: string  
  TxResponse:
    type: object
    properties:
      height:
        type: string
      txhash:
        type: string
      codespace:
        type: string
      code:
        type: number
      data:
        type: string
      raw_log:
        type: string
      logs:
        type: array
        items: 
          $ref: '#/definitions/Log'
      info:
        type: string
      gas_wanted:
        type: string
      gas_used:
        type: string
      tx:
        type: object
        properties:
          "@type":
            type: string
          body:
            type: object
            properties:
              messages:
                type: array
                items:
                  $ref: '#/definitions/TypeValue'
              memo: 
                type: string
              timeout_height:
                type: string
              extension_options: 
                type: array
                items: 
                  $ref: '#/definitions/TypeValue'
              non_critical_extension_options:
                type: array
                items: 
                  $ref: '#/definitions/TypeValue'
          auth_info:
            type: object
            properties:
              signer_infos:
                type: array
                items:
                  $ref: '#/definitions/Signer' 
              fee:
                type: object
                properties:
                  amount:
                    type: array
                    items: 
                      $ref: '#/definitions/Amount'
                  gas_limit:
                    type: string
                  payer:
                    type: string
                  granter:
                    type: string 
          signature:
            type: array
            items:
              type: string                 
      timestamp:
        type: string        
  Amount:
    type: object
    properties:
      denom:
        type: string
      amount: 
        type: string   
  Signer:
    type: object
    properties: 
      public_key:
        type: object
        properties:
          "@type":
            type: string
          key:
            type: string  
      mode_info:
        type: object
        properties:
          single:
            type: object
            properties: 
              mode:
                type: string
          sequence:
            type: string       
  Log:
    type: object
    properties:
      msg_index: 
        type: number
      log:
        type: string
      events:
        type: array
        items:
          type: object
          properties:
            type:
              type: string
            attributes:
              type: array
              items:
                type: object
                properties:
                  key:
                    type: string
                  value:
                    type: string
  TypeValue:
    type: object
    properties:
      type_url:
        type: string
      value:
        type: string